#2 Moved DisposeAssetAction to a separate child package (.actions).
