+++ airport/airport-web-server/src/main/java/jhou/webapp/WebUiConfig.java
@@ -4,7 +4,7 @@ import org.apache.commons.lang.StringUtils;
 
 import jhou.config.personnel.PersonWebUiConfig;
 import jhou.webapp.config.asset.AssetWebUiConfig;
-import jhou.webapp.config.asset.DisposeAssetActionWebUiConfig;
+import jhou.webapp.config.asset.actions.DisposeAssetActionWebUiConfig;
 import jhou.webapp.config.tablecodes.AssetClassWebUiConfig;
 import ua.com.fielden.platform.basic.config.Workflows;
 import ua.com.fielden.platform.web.app.config.IWebUiBuilder;
